# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: OpenAI AI chat, image, and audio operations
  - name: pets
    description: Pet management
  - name: posts
    description: Community posts
  - name: profile
    description: User profile
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /openai/generate-image:
    post:
      operationId: generateOpenaiImage
      tags: [openai]
      summary: Generate an AI styled pet image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOpenaiImageBody"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateOpenaiImageResponse"
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all community posts
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new community post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like or unlike a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated like count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /pets:
    get:
      operationId: listPets
      tags: [pets]
      summary: List user's pets
      responses:
        "200":
          description: List of pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pet"
    post:
      operationId: createPet
      tags: [pets]
      summary: Add a new pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePetBody"
      responses:
        "201":
          description: Created pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
  /pets/{id}:
    delete:
      operationId: deletePet
      tags: [pets]
      summary: Delete a pet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    GenerateOpenaiImageBody:
      type: object
      properties:
        prompt:
          type: string
        size:
          type: string
          enum: ["1024x1024", "512x512", "256x256"]
      required:
        - prompt
    GenerateOpenaiImageResponse:
      type: object
      properties:
        b64_json:
          type: string
      required:
        - b64_json
    Post:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        userAvatar:
          type: string
        petName:
          type: string
        petType:
          type: string
        imageData:
          type: string
        style:
          type: string
        caption:
          type: string
        likes:
          type: integer
        likedByMe:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - petName
        - petType
        - imageData
        - style
        - caption
        - likes
        - likedByMe
        - createdAt
    CreatePostBody:
      type: object
      properties:
        petName:
          type: string
        petType:
          type: string
        imageData:
          type: string
        style:
          type: string
        caption:
          type: string
        userName:
          type: string
        userAvatar:
          type: string
      required:
        - petName
        - petType
        - imageData
        - style
        - caption
        - userName
    Pet:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        breed:
          type: string
        imageData:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - createdAt
    CreatePetBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        breed:
          type: string
        imageData:
          type: string
      required:
        - name
        - type
    Profile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        avatarData:
          type: string
        postsCount:
          type: integer
        followersCount:
          type: integer
        followingCount:
          type: integer
      required:
        - id
        - username
        - displayName
        - postsCount
        - followersCount
        - followingCount
    UpdateProfileBody:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        avatarData:
          type: string