# lib/api-zod/src/generated/types/createPetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePetBody {
  name: string;
  type: string;
  breed?: string;
  imageData?: string;
}