# lib/api-zod/src/generated/types/createPostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePostBody {
  petName: string;
  petType: string;
  imageData: string;
  style: string;
  caption: string;
  userName: string;
  userAvatar?: string;
}