# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Post {
  id: string;
  userId: string;
  userName: string;
  userAvatar?: string;
  petName: string;
  petType: string;
  imageData: string;
  style: string;
  caption: string;
  likes: number;
  likedByMe: boolean;
  createdAt: Date;
}