# lib/api-zod/src/generated/types/updateProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileBody {
  username?: string;
  displayName?: string;
  bio?: string;
  avatarData?: string;
}